pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ebiosrm-generator"
version = "2.0.0"
description = "EBIOS Risk Manager generator - modular version"
authors = [{name = "EBIOS Team", email = "team@ebios.example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "coverage>=7.0",
]

[project.scripts]
ebiosrm = "ebiosrm_core.cli:app"

[tool.setuptools.packages.find]
where = [""]
include = ["ebiosrm_core*", "cli.py"]
exclude = ["config*", "tests*", "build*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
